Move dialog from internal to public
